backend/pyproject.toml
[tool.poetry]
name = "quick_wallet"
version = "0.0.1"
description = ""
authors = ["Ivan Morgun <i.morgun@g.nsu.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
fastapi-pagination = "^0.11.0"
uvicorn = "^0.20.0"
smsaero-api = "^2.1.1"
SQLAlchemy = "^1.4.44"
singleton-decorator = "^1.0.0"
pytz = "^2022.6"
asyncpg = "^0.27.0"
python-dotenv = "^0.21.0"
PyJWT = "^2.6.0"
isort = "^5.10.1"
black = "^22.10.0"
orjson = "^3.8.3"
decorator = "^5.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
